refactor: split sql builder code
