+++ src/utils/SQLBuilder/GroupExpression.ts
@@ -1,0 +1,12 @@
+import { RawQuery } from './core/RawQuery';
+import { QuerySegment } from '.';
+
+export class GroupExpression extends RawQuery {
+	constructor(...query: (QuerySegment | string)[]) {
+		if (query.length === 0) {
+			super();
+		} else {
+			super('(', ...query, ')');
+		}
+	}
+}
