+++ src/utils/SQLBuilder/SQLBuilder.test.ts
@@ -1,17 +1,15 @@
-import {
-	ConditionClause,
-	GroupExpression,
-	Limit,
-	ListExpression,
-	Placeholder,
-	QueryConstrictor,
-	WhereClause,
-} from './QueryBuilder';
+import { ConditionClause } from './ConditionClause';
+import { PreparedValue } from './core/PreparedValue';
+import { GroupExpression } from './GroupExpression';
+import { LimitClause } from './LimitClause';
+import { SetExpression } from './SetExpression';
+import { QueryConstructor } from './utils/QueryConstructor';
+import { WhereClause } from './WhereClause';
 
 describe('Primitives', () => {
 	test('Query constructor able to add queries one by one', () => {
-		const query = new QueryConstrictor();
-		query.add('SELECT * FROM foo WHERE foo=', query.getPlaceholder(1));
+		const query = new QueryConstructor();
+		query.add('SELECT * FROM foo WHERE foo=', query.value(1));
 		query.add(' LIMIT 2');
 
 		expect(query.toSQL()).toEqual({
@@ -21,10 +19,10 @@ describe('Primitives', () => {
 	});
 
 	test('Query constructors may be nested', () => {
-		const query1 = new QueryConstrictor();
+		const query1 = new QueryConstructor();
 
-		const query2 = new QueryConstrictor();
-		query2.add('(SELECT id FROM bar WHERE x > ', query2.getPlaceholder(100), ')');
+		const query2 = new QueryConstructor();
+		query2.add('(SELECT id FROM bar WHERE x > ', query2.value(100), ')');
 		query1.add('SELECT * FROM foo WHERE foo IN ', query2);
 
 		expect(query1.toSQL()).toEqual({
@@ -34,11 +32,11 @@ describe('Primitives', () => {
 	});
 
 	test('Query constructors may be nested with no introduce a variables', () => {
-		const query = new QueryConstrictor(
+		const query = new QueryConstructor(
 			'SELECT * FROM foo WHERE foo IN ',
-			new QueryConstrictor(
+			new QueryConstructor(
 				'(SELECT id FROM bar WHERE x > ',
-				new Placeholder(100),
+				new PreparedValue(100),
 				')',
 			),
 		);
@@ -62,9 +60,9 @@ describe('Basic clauses', () => {
 		test('Nested group expression yields correct query', () => {
 			expect(
 				new GroupExpression(
-					new GroupExpression(new Placeholder(1)),
+					new GroupExpression(new PreparedValue(1)),
 					' OR ',
-					new GroupExpression(new Placeholder(2)),
+					new GroupExpression(new PreparedValue(2)),
 				).toSQL(),
 			).toEqual({
 				sql: '((?) OR (?))',
@@ -73,16 +71,16 @@ describe('Basic clauses', () => {
 		});
 	});
 
-	describe('List expression', () => {
-		test('List with literals', () => {
-			expect(new ListExpression('foo', 'bar', 'baz').toSQL()).toEqual({
+	describe('Set expression', () => {
+		test('Set with literals', () => {
+			expect(new SetExpression('foo', 'bar', 'baz').toSQL()).toEqual({
 				sql: '(foo,bar,baz)',
 				bindings: [],
 			});
 		});
-		test('Nested lists', () => {
+		test('Nested sets', () => {
 			expect(
-				new ListExpression('foo', new ListExpression('bar', 'baz')).toSQL(),
+				new SetExpression('foo', new SetExpression('bar', 'baz')).toSQL(),
 			).toEqual({
 				sql: '(foo,(bar,baz))',
 				bindings: [],
@@ -92,19 +90,19 @@ describe('Basic clauses', () => {
 
 	describe('Limit expression', () => {
 		test('Limit', () => {
-			expect(new Limit({ limit: 10 }).toSQL()).toEqual({
+			expect(new LimitClause({ limit: 10 }).toSQL()).toEqual({
 				sql: 'LIMIT 10',
 				bindings: [],
 			});
 		});
 		test('Offset', () => {
-			expect(new Limit({ offset: 20 }).toSQL()).toEqual({
+			expect(new LimitClause({ offset: 20 }).toSQL()).toEqual({
 				sql: 'OFFSET 20',
 				bindings: [],
 			});
 		});
 		test('Limit and offset', () => {
-			expect(new Limit({ limit: 10, offset: 20 }).toSQL()).toEqual({
+			expect(new LimitClause({ limit: 10, offset: 20 }).toSQL()).toEqual({
 				sql: 'LIMIT 10 OFFSET 20',
 				bindings: [],
 			});
@@ -123,14 +121,16 @@ describe('Basic clauses', () => {
 		});
 
 		test('Single condition yields just a literal', () => {
-			expect(new ConditionClause().and('x > ', new Placeholder(0)).toSQL()).toEqual(
-				{
-					sql: 'x > ?',
-					bindings: [0],
-				},
-			);
+			expect(
+				new ConditionClause().and('x > ', new PreparedValue(0)).toSQL(),
+			).toEqual({
+				sql: 'x > ?',
+				bindings: [0],
+			});
 
-			expect(new ConditionClause().or('x > ', new Placeholder(0)).toSQL()).toEqual({
+			expect(
+				new ConditionClause().or('x > ', new PreparedValue(0)).toSQL(),
+			).toEqual({
 				sql: 'x > ?',
 				bindings: [0],
 			});
@@ -139,8 +139,8 @@ describe('Basic clauses', () => {
 		test('Trivial condition expression yields sql expression', () => {
 			expect(
 				new ConditionClause()
-					.and('x > ', new Placeholder(0))
-					.or('y < ', new Placeholder(1))
+					.and('x > ', new PreparedValue(0))
+					.or('y < ', new PreparedValue(1))
 					.toSQL(),
 			).toEqual({
 				sql: 'x > ? OR y < ?',
@@ -149,15 +149,15 @@ describe('Basic clauses', () => {
 		});
 
 		test('Complex condition expression consider a grouping', () => {
-			const query = new QueryConstrictor(
+			const query = new QueryConstructor(
 				'SELECT * FROM foo WHERE ',
 				new ConditionClause()
-					.and('x > ', new Placeholder(0))
+					.and('x > ', new PreparedValue(0))
 					.or(
 						new GroupExpression(
 							new ConditionClause()
-								.and('y=', new Placeholder(1))
-								.and('z=', new Placeholder(2)),
+								.and('y=', new PreparedValue(1))
+								.and('z=', new PreparedValue(2)),
 						),
 					),
 			);
@@ -172,16 +172,16 @@ describe('Basic clauses', () => {
 	describe('Where clause', () => {
 		test('Where clause may be filled after join', () => {
 			const where = new WhereClause();
-			const query = new QueryConstrictor('SELECT * FROM foo ', where);
+			const query = new QueryConstructor('SELECT * FROM foo ', where);
 
 			// Fill where after build a query object
 			where
-				.and('x > ', new Placeholder(0))
+				.and('x > ', new PreparedValue(0))
 				.or(
 					new GroupExpression(
 						new ConditionClause()
-							.and('y=', new Placeholder(1))
-							.and('z=', new Placeholder(2)),
+							.and('y=', new PreparedValue(1))
+							.and('z=', new PreparedValue(2)),
 					),
 				);
