+++ src/utils/SQLBuilder/core/PreparedValue.ts
@@ -1,0 +1,3 @@
+import { RawValue } from './RawValue';
+
+export class PreparedValue extends RawValue {}
