+++ src/utils/SQLBuilder/core/RawValue.ts
@@ -1,0 +1,10 @@
+export class RawValue {
+	protected readonly value;
+	constructor(value: string | number) {
+		this.value = value;
+	}
+
+	public getValue = () => {
+		return this.value;
+	};
+}
